backend/requirements.txt
# College List AI - Backend Dependencies

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication
PyJWT[crypto]>=2.8.0

# Database
supabase>=2.0.0
sqlmodel>=0.0.22
asyncpg>=0.29.0
alembic>=1.14.0
greenlet>=3.0.0

# AI/ML
google-genai>=1.0.0
langgraph>=1.0.0

# Payments
stripe>=8.0.0

# Utilities
python-dotenv>=1.0.0
slowapi>=0.1.9

# Development
httpx>=0.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
